move migration logic and interval from main process to render process, because session state and a VCS instance dont work in the main process yet
