+++ packages/insomnia-app/app/sync/vcs/migrate-collections.ts
@@ -22,7 +22,7 @@ export const logCollectionMovedToSpace = (collection: Workspace, remoteSpace: Re
 };
 
 export const migrateCollectionsIntoRemoteSpace = async (vcs: VCS) => {
-  console.log('[sync] checking for collections which need to be migrated');
+  console.log('[sync] checking for collections which need to be moved into a remote space');
 
   // If not logged in, exit
   if (!isLoggedIn()) {
