+++ packages/insomnia-app/app/ui/containers/app-hooks.tsx
@@ -1,0 +1,9 @@
+import { FC } from 'react';
+
+import { useSyncMigration } from '../hooks/use-sync-migration';
+
+export const AppHooks: FC = () => {
+  useSyncMigration();
+
+  return null;
+};
