+++ packages/insomnia-app/app/ui/hooks/effect-helpers.ts
@@ -1,0 +1,6 @@
+export const invokeAsyncSynchronously = (callback: () => Promise<void>) => {
+  async function func() {
+    await callback();
+  }
+  func();
+};
